backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
celery==5.4.0
redis==5.2.1
boto3==1.35.86
python-docx==1.1.2
openpyxl==3.1.5
pypdf==5.1.0
openai==1.58.1
